# system/nodejs-server-server/api/swagger.yaml
---
swagger: "2.0"
info:
  version: "0.0.1"
  title: "Siren"
host: "localhost:8080"
basePath: "/fandango/v0.1/siren"
schemes:
- "http"
paths:
  /FindSimilarClaims:
    post:
      summary: "Find claims similar to existing claim, and similar to optional topics"
      description: "Response will return top 5 Claims closest to user provided criteria.\
        \ Each claim in result will contain the claim itself, and a list of all claim\
        \ reviews of that claim."
      operationId: "findSimilarClaims"
      parameters:
      - in: "body"
        name: "info"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Claim_Request_Interface"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Claim_Response_Interface"
      x-swagger-router-controller: "Default"
  /FindSimilarArticles:
    post:
      summary: "Find articles similar to an existing article"
      description: "Response will return top 5 Articles closest to user provided criteria."
      operationId: "findSimilarArticles"
      parameters:
      - in: "body"
        name: "info"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Article_Request_Interface"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Article_Response_Interface"
      x-swagger-router-controller: "Default"
  /FindFact:
    post:
      summary: "Find Facts similar to text input, specific category, and similar to\
        \ optional topics"
      description: "Response will return top 5 Fact documents closest to user provided\
        \ criteria. A Fact is an Open Data creative work containing data, published\
        \ by a recognized institution to make information available publicly."
      operationId: "findFact"
      parameters:
      - in: "body"
        name: "info"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Fact_Request_Interface"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Fact_Response_Interface"
      x-swagger-router-controller: "Default"
definitions:
  Claim_Request_Interface:
    type: "object"
    required:
    - "identifier"
    properties:
      identifier:
        type: "integer"
      topics:
        type: "array"
        items:
          type: "string"
  Claim_Response_Interface:
    type: "object"
    required:
    - "identifier"
    - "results"
    properties:
      identifier:
        type: "integer"
      topics:
        type: "array"
        items:
          type: "string"
      results:
        type: "array"
        items:
          $ref: "#/definitions/Claim_Response_Interface_results"
    example:
      identifier: 0
      topics:
      - "topics"
      - "topics"
      results:
      - claim:
          firstAppearance: 2
          identifier: 7
          claimReviews:
          - datePublished: "datePublished"
            identifier: 2
            reviewAspect: "reviewAspect"
            claimReviewed: "claimReviewed"
            dateCreated: "dateCreated"
            references:
            - 1
            - 1
            reviewBody: "reviewBody"
            itemReviewed: 7
            dateModified: "dateModified"
            aggregateRating: 4.1456080298839363962315474054776132106781005859375
          - datePublished: "datePublished"
            identifier: 2
            reviewAspect: "reviewAspect"
            claimReviewed: "claimReviewed"
            dateCreated: "dateCreated"
            references:
            - 1
            - 1
            reviewBody: "reviewBody"
            itemReviewed: 7
            dateModified: "dateModified"
            aggregateRating: 4.1456080298839363962315474054776132106781005859375
          about:
          - "about"
          - "about"
          calculatedRatingDetail:
            authorRating: 5.63737665663332876420099637471139430999755859375
            publishRating: 5.962133916683182377482808078639209270477294921875
          dateModified: "dateModified"
          calculatedRating: 1.46581298050294517310021547018550336360931396484375
          datePublished: "datePublished"
          appearance:
          - 6
          - 6
          dateCreated: "dateCreated"
          mentions:
          - 9
          - 9
          text: "text"
          possiblyRelatesTo:
          - 3
          - 3
      - claim:
          firstAppearance: 2
          identifier: 7
          claimReviews:
          - datePublished: "datePublished"
            identifier: 2
            reviewAspect: "reviewAspect"
            claimReviewed: "claimReviewed"
            dateCreated: "dateCreated"
            references:
            - 1
            - 1
            reviewBody: "reviewBody"
            itemReviewed: 7
            dateModified: "dateModified"
            aggregateRating: 4.1456080298839363962315474054776132106781005859375
          - datePublished: "datePublished"
            identifier: 2
            reviewAspect: "reviewAspect"
            claimReviewed: "claimReviewed"
            dateCreated: "dateCreated"
            references:
            - 1
            - 1
            reviewBody: "reviewBody"
            itemReviewed: 7
            dateModified: "dateModified"
            aggregateRating: 4.1456080298839363962315474054776132106781005859375
          about:
          - "about"
          - "about"
          calculatedRatingDetail:
            authorRating: 5.63737665663332876420099637471139430999755859375
            publishRating: 5.962133916683182377482808078639209270477294921875
          dateModified: "dateModified"
          calculatedRating: 1.46581298050294517310021547018550336360931396484375
          datePublished: "datePublished"
          appearance:
          - 6
          - 6
          dateCreated: "dateCreated"
          mentions:
          - 9
          - 9
          text: "text"
          possiblyRelatesTo:
          - 3
          - 3
  Article_Request_Interface:
    type: "object"
    required:
    - "identifier"
    properties:
      identifier:
        type: "integer"
  Article_Response_Interface:
    type: "object"
    required:
    - "identifier"
    - "results"
    properties:
      identifier:
        type: "integer"
      results:
        type: "array"
        items:
          $ref: "#/definitions/Article_Response_Interface_results"
    example:
      identifier: 0
      results:
      - article:
          identifier: 6
          articleBody: "articleBody"
          author:
          - identifier: 1
            nationality: "nationality"
            gender: "gender"
            affiliation:
            - 5
            - 5
            bias: "bias"
            jobTitle: "jobTitle"
            name: "name"
            url: "url"
          - identifier: 1
            nationality: "nationality"
            gender: "gender"
            affiliation:
            - 5
            - 5
            bias: "bias"
            jobTitle: "jobTitle"
            name: "name"
            url: "url"
          about:
          - "about"
          - "about"
          dateModified: "dateModified"
          calculatedRatingDetail:
            authorRating: 5.63737665663332876420099637471139430999755859375
            publishRating: 5.962133916683182377482808078639209270477294921875
          datePublished: "datePublished"
          calculatedRating: 3.61607674925191080461672754609026014804840087890625
          contains:
          - 9
          - 9
          dateCreated: "dateCreated"
          mentions:
          - 7
          - 7
          publisher:
          - identifier: 5
            nationality: "nationality"
            bias: "bias"
            name: "name"
            parentOrganization:
            - 2
            - 2
            url: "url"
          - identifier: 5
            nationality: "nationality"
            bias: "bias"
            name: "name"
            parentOrganization:
            - 2
            - 2
            url: "url"
          headline: "headline"
      - article:
          identifier: 6
          articleBody: "articleBody"
          author:
          - identifier: 1
            nationality: "nationality"
            gender: "gender"
            affiliation:
            - 5
            - 5
            bias: "bias"
            jobTitle: "jobTitle"
            name: "name"
            url: "url"
          - identifier: 1
            nationality: "nationality"
            gender: "gender"
            affiliation:
            - 5
            - 5
            bias: "bias"
            jobTitle: "jobTitle"
            name: "name"
            url: "url"
          about:
          - "about"
          - "about"
          dateModified: "dateModified"
          calculatedRatingDetail:
            authorRating: 5.63737665663332876420099637471139430999755859375
            publishRating: 5.962133916683182377482808078639209270477294921875
          datePublished: "datePublished"
          calculatedRating: 3.61607674925191080461672754609026014804840087890625
          contains:
          - 9
          - 9
          dateCreated: "dateCreated"
          mentions:
          - 7
          - 7
          publisher:
          - identifier: 5
            nationality: "nationality"
            bias: "bias"
            name: "name"
            parentOrganization:
            - 2
            - 2
            url: "url"
          - identifier: 5
            nationality: "nationality"
            bias: "bias"
            name: "name"
            parentOrganization:
            - 2
            - 2
            url: "url"
          headline: "headline"
  Fact_Request_Interface:
    type: "object"
    required:
    - "category"
    - "text"
    properties:
      text:
        type: "string"
      category:
        type: "string"
      topics:
        type: "array"
        items:
          type: "string"
  Fact_Response_Interface:
    type: "object"
    required:
    - "category"
    - "results"
    - "text"
    properties:
      text:
        type: "string"
      category:
        type: "string"
      topics:
        type: "array"
        items:
          type: "string"
      results:
        type: "array"
        items:
          $ref: "#/definitions/Fact_Response_Interface_results"
    example:
      topics:
      - "topics"
      - "topics"
      text: "text"
      category: "category"
      results:
      - fact:
          datePublished: "datePublished"
          identifier: 0
          dateCreated: "dateCreated"
          temporalCoverageStart: "temporalCoverageStart"
          dateMtemporalCoverageEnd: "dateMtemporalCoverageEnd"
          mentions:
          - 6
          - 6
          name: "name"
          about:
          - "about"
          - "about"
          dateModified: "dateModified"
          text: "text"
          spatialCoverage: "spatialCoverage"
          url: "url"
      - fact:
          datePublished: "datePublished"
          identifier: 0
          dateCreated: "dateCreated"
          temporalCoverageStart: "temporalCoverageStart"
          dateMtemporalCoverageEnd: "dateMtemporalCoverageEnd"
          mentions:
          - 6
          - 6
          name: "name"
          about:
          - "about"
          - "about"
          dateModified: "dateModified"
          text: "text"
          spatialCoverage: "spatialCoverage"
          url: "url"
  Claim_Response_Interface_claim_calculatedRatingDetail:
    properties:
      publishRating:
        type: "number"
      authorRating:
        type: "number"
    example:
      authorRating: 5.63737665663332876420099637471139430999755859375
      publishRating: 5.962133916683182377482808078639209270477294921875
  Claim_Response_Interface_claim_claimReviews:
    properties:
      identifier:
        type: "integer"
      claimReviewed:
        type: "string"
      reviewAspect:
        type: "string"
      reviewBody:
        type: "string"
      dateCreated:
        type: "string"
      dateModified:
        type: "string"
      datePublished:
        type: "string"
      aggregateRating:
        type: "number"
      itemReviewed:
        type: "integer"
      references:
        type: "array"
        items:
          type: "integer"
    example:
      datePublished: "datePublished"
      identifier: 2
      reviewAspect: "reviewAspect"
      claimReviewed: "claimReviewed"
      dateCreated: "dateCreated"
      references:
      - 1
      - 1
      reviewBody: "reviewBody"
      itemReviewed: 7
      dateModified: "dateModified"
      aggregateRating: 4.1456080298839363962315474054776132106781005859375
  Claim_Response_Interface_claim:
    properties:
      about:
        type: "array"
        items:
          type: "string"
      appearance:
        type: "array"
        items:
          type: "integer"
      calculatedRating:
        type: "number"
      calculatedRatingDetail:
        $ref: "#/definitions/Claim_Response_Interface_claim_calculatedRatingDetail"
      dateCreated:
        type: "string"
      dateModified:
        type: "string"
      datePublished:
        type: "string"
      firstAppearance:
        type: "integer"
      identifier:
        type: "integer"
      mentions:
        type: "array"
        items:
          type: "integer"
      possiblyRelatesTo:
        type: "array"
        items:
          type: "integer"
      text:
        type: "string"
      claimReviews:
        type: "array"
        items:
          $ref: "#/definitions/Claim_Response_Interface_claim_claimReviews"
    example:
      firstAppearance: 2
      identifier: 7
      claimReviews:
      - datePublished: "datePublished"
        identifier: 2
        reviewAspect: "reviewAspect"
        claimReviewed: "claimReviewed"
        dateCreated: "dateCreated"
        references:
        - 1
        - 1
        reviewBody: "reviewBody"
        itemReviewed: 7
        dateModified: "dateModified"
        aggregateRating: 4.1456080298839363962315474054776132106781005859375
      - datePublished: "datePublished"
        identifier: 2
        reviewAspect: "reviewAspect"
        claimReviewed: "claimReviewed"
        dateCreated: "dateCreated"
        references:
        - 1
        - 1
        reviewBody: "reviewBody"
        itemReviewed: 7
        dateModified: "dateModified"
        aggregateRating: 4.1456080298839363962315474054776132106781005859375
      about:
      - "about"
      - "about"
      calculatedRatingDetail:
        authorRating: 5.63737665663332876420099637471139430999755859375
        publishRating: 5.962133916683182377482808078639209270477294921875
      dateModified: "dateModified"
      calculatedRating: 1.46581298050294517310021547018550336360931396484375
      datePublished: "datePublished"
      appearance:
      - 6
      - 6
      dateCreated: "dateCreated"
      mentions:
      - 9
      - 9
      text: "text"
      possiblyRelatesTo:
      - 3
      - 3
  Claim_Response_Interface_results:
    properties:
      claim:
        $ref: "#/definitions/Claim_Response_Interface_claim"
    example:
      claim:
        firstAppearance: 2
        identifier: 7
        claimReviews:
        - datePublished: "datePublished"
          identifier: 2
          reviewAspect: "reviewAspect"
          claimReviewed: "claimReviewed"
          dateCreated: "dateCreated"
          references:
          - 1
          - 1
          reviewBody: "reviewBody"
          itemReviewed: 7
          dateModified: "dateModified"
          aggregateRating: 4.1456080298839363962315474054776132106781005859375
        - datePublished: "datePublished"
          identifier: 2
          reviewAspect: "reviewAspect"
          claimReviewed: "claimReviewed"
          dateCreated: "dateCreated"
          references:
          - 1
          - 1
          reviewBody: "reviewBody"
          itemReviewed: 7
          dateModified: "dateModified"
          aggregateRating: 4.1456080298839363962315474054776132106781005859375
        about:
        - "about"
        - "about"
        calculatedRatingDetail:
          authorRating: 5.63737665663332876420099637471139430999755859375
          publishRating: 5.962133916683182377482808078639209270477294921875
        dateModified: "dateModified"
        calculatedRating: 1.46581298050294517310021547018550336360931396484375
        datePublished: "datePublished"
        appearance:
        - 6
        - 6
        dateCreated: "dateCreated"
        mentions:
        - 9
        - 9
        text: "text"
        possiblyRelatesTo:
        - 3
        - 3
  Article_Response_Interface_article_author:
    properties:
      identifier:
        type: "integer"
      name:
        type: "string"
      url:
        type: "string"
      nationality:
        type: "string"
      bias:
        type: "string"
      jobTitle:
        type: "string"
      gender:
        type: "string"
      affiliation:
        type: "array"
        items:
          type: "integer"
    example:
      identifier: 1
      nationality: "nationality"
      gender: "gender"
      affiliation:
      - 5
      - 5
      bias: "bias"
      jobTitle: "jobTitle"
      name: "name"
      url: "url"
  Article_Response_Interface_article_publisher:
    properties:
      identifier:
        type: "integer"
      name:
        type: "string"
      url:
        type: "string"
      nationality:
        type: "string"
      bias:
        type: "string"
      parentOrganization:
        type: "array"
        items:
          type: "integer"
    example:
      identifier: 5
      nationality: "nationality"
      bias: "bias"
      name: "name"
      parentOrganization:
      - 2
      - 2
      url: "url"
  Article_Response_Interface_article:
    properties:
      identifier:
        type: "integer"
      headline:
        type: "string"
      articleBody:
        type: "string"
      dateCreated:
        type: "string"
      dateModified:
        type: "string"
      datePublished:
        type: "string"
      author:
        type: "array"
        items:
          $ref: "#/definitions/Article_Response_Interface_article_author"
      publisher:
        type: "array"
        items:
          $ref: "#/definitions/Article_Response_Interface_article_publisher"
      mentions:
        type: "array"
        items:
          type: "integer"
      contains:
        type: "array"
        items:
          type: "integer"
      about:
        type: "array"
        items:
          type: "string"
      calculatedRating:
        type: "number"
      calculatedRatingDetail:
        $ref: "#/definitions/Claim_Response_Interface_claim_calculatedRatingDetail"
    example:
      identifier: 6
      articleBody: "articleBody"
      author:
      - identifier: 1
        nationality: "nationality"
        gender: "gender"
        affiliation:
        - 5
        - 5
        bias: "bias"
        jobTitle: "jobTitle"
        name: "name"
        url: "url"
      - identifier: 1
        nationality: "nationality"
        gender: "gender"
        affiliation:
        - 5
        - 5
        bias: "bias"
        jobTitle: "jobTitle"
        name: "name"
        url: "url"
      about:
      - "about"
      - "about"
      dateModified: "dateModified"
      calculatedRatingDetail:
        authorRating: 5.63737665663332876420099637471139430999755859375
        publishRating: 5.962133916683182377482808078639209270477294921875
      datePublished: "datePublished"
      calculatedRating: 3.61607674925191080461672754609026014804840087890625
      contains:
      - 9
      - 9
      dateCreated: "dateCreated"
      mentions:
      - 7
      - 7
      publisher:
      - identifier: 5
        nationality: "nationality"
        bias: "bias"
        name: "name"
        parentOrganization:
        - 2
        - 2
        url: "url"
      - identifier: 5
        nationality: "nationality"
        bias: "bias"
        name: "name"
        parentOrganization:
        - 2
        - 2
        url: "url"
      headline: "headline"
  Article_Response_Interface_results:
    properties:
      article:
        $ref: "#/definitions/Article_Response_Interface_article"
    example:
      article:
        identifier: 6
        articleBody: "articleBody"
        author:
        - identifier: 1
          nationality: "nationality"
          gender: "gender"
          affiliation:
          - 5
          - 5
          bias: "bias"
          jobTitle: "jobTitle"
          name: "name"
          url: "url"
        - identifier: 1
          nationality: "nationality"
          gender: "gender"
          affiliation:
          - 5
          - 5
          bias: "bias"
          jobTitle: "jobTitle"
          name: "name"
          url: "url"
        about:
        - "about"
        - "about"
        dateModified: "dateModified"
        calculatedRatingDetail:
          authorRating: 5.63737665663332876420099637471139430999755859375
          publishRating: 5.962133916683182377482808078639209270477294921875
        datePublished: "datePublished"
        calculatedRating: 3.61607674925191080461672754609026014804840087890625
        contains:
        - 9
        - 9
        dateCreated: "dateCreated"
        mentions:
        - 7
        - 7
        publisher:
        - identifier: 5
          nationality: "nationality"
          bias: "bias"
          name: "name"
          parentOrganization:
          - 2
          - 2
          url: "url"
        - identifier: 5
          nationality: "nationality"
          bias: "bias"
          name: "name"
          parentOrganization:
          - 2
          - 2
          url: "url"
        headline: "headline"
  Fact_Response_Interface_fact:
    properties:
      identifier:
        type: "integer"
      name:
        type: "string"
      text:
        type: "string"
      url:
        type: "string"
      about:
        type: "array"
        items:
          type: "string"
      mentions:
        type: "array"
        items:
          type: "integer"
      dateCreated:
        type: "string"
      dateModified:
        type: "string"
      datePublished:
        type: "string"
      temporalCoverageStart:
        type: "string"
      dateMtemporalCoverageEnd:
        type: "string"
      spatialCoverage:
        type: "string"
    example:
      datePublished: "datePublished"
      identifier: 0
      dateCreated: "dateCreated"
      temporalCoverageStart: "temporalCoverageStart"
      dateMtemporalCoverageEnd: "dateMtemporalCoverageEnd"
      mentions:
      - 6
      - 6
      name: "name"
      about:
      - "about"
      - "about"
      dateModified: "dateModified"
      text: "text"
      spatialCoverage: "spatialCoverage"
      url: "url"
  Fact_Response_Interface_results:
    properties:
      fact:
        $ref: "#/definitions/Fact_Response_Interface_fact"
    example:
      fact:
        datePublished: "datePublished"
        identifier: 0
        dateCreated: "dateCreated"
        temporalCoverageStart: "temporalCoverageStart"
        dateMtemporalCoverageEnd: "dateMtemporalCoverageEnd"
        mentions:
        - 6
        - 6
        name: "name"
        about:
        - "about"
        - "about"
        dateModified: "dateModified"
        text: "text"
        spatialCoverage: "spatialCoverage"
        url: "url"